day13@AoC24: P2: Success

+++ src/main/kotlin/com/xing/aoc24/day13/Machine.kt
@@ -1,14 +1,14 @@
 package com.xing.aoc24.day13
 
-data class Machine(val aX: Int, val aY: Int, val bX: Int, val bY: Int, val prizeX: Int, val prizeY: Int) {
-    fun cost(): Int? {
+data class Machine(val aX: Int, val aY: Int, val bX: Int, val bY: Int, val prizeX: Long, val prizeY: Long) {
+    fun cost(): Long? {
         val b1 = prizeY * aX - aY * prizeX
         val b2 = aX * bY - aY * bX
-        if (b1 % b2 != 0) return null
+        if (b1 % b2 != 0L) return null
         val b = b1 / b2
         val a1 = prizeX - b * bX
         val a2 = aX
-        if (a1 % a2 != 0) return null
+        if (a1 % a2 != 0L) return null
         val a = a1 / a2
         return a * 3 + b
     }
